+++ qcmanybody/__init__.py
@@ -1,5 +1,8 @@
 from importlib.metadata import version
 
+__version__ = version("qcmanybody")
+del version
+
 # isort: off
 from .core import ManyBodyCalculator  # legacy near-alias to ManyBodyCore
 from .core import ManyBodyCore
@@ -8,6 +11,3 @@ from .computer import ManyBodyComputer
 # isort: on
 from .models import BsseEnum
 from .utils import delabeler, labeler, resize_gradient, resize_hessian
-
-__version__ = version("qcmanybody")
-del version